Move SQLite handler to example only
